+++ addon/components/decision-plugin/decision-plugin-card.ts
@@ -1,9 +1,13 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
-import { insertTitle } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/decision-plugin/commands';
+import {
+  insertMotivation,
+  insertArticleContainer,
+} from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/decision-plugin/commands';
 import { SayController } from '@lblod/ember-rdfa-editor';
 import { inject as service } from '@ember/service';
 import IntlService from 'ember-intl/services/intl';
+import { VALIDATION_KEY } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/validation';
 
 type Args = {
   controller: SayController;
@@ -24,26 +28,41 @@ export default class DecisionPluginCard extends Component<Args> {
   }
 
   @action
-  insertTitle() {
-    this.controller.doCommand(
-      insertTitle({
-        placeholderText: this.intl.t('besluit-plugin.insert.decision-title'),
-      }),
-      {
-        view: this.controller.mainEditorView,
-      }
-    );
+  insertMotivation() {
+    this.controller.doCommand(insertMotivation({ intl: this.intl }), {
+      view: this.controller.mainEditorView,
+    });
     this.focus();
   }
 
-  get canInsertTitle() {
-    return this.controller.checkCommand(
-      insertTitle({
-        placeholderText: this.intl.t('besluit-plugin.insert.decision-title'),
-      }),
-      {
-        view: this.controller.mainEditorView,
+  get canInsertMotivation() {
+    return this.controller.checkCommand(insertMotivation({ intl: this.intl }), {
+      view: this.controller.mainEditorView,
+    });
+  }
+
+  @action
+  insertArticleBlock() {
+    this.controller.doCommand(insertArticleContainer({ intl: this.intl }), {
+      view: this.controller.mainEditorView,
+    });
+    this.focus();
+  }
+
+  get missingArticleBlock() {
+    const state = VALIDATION_KEY.getState(this.controller.activeEditorState);
+    if (state) {
+      let missingArticleBlock = false;
+      const results = state.report.results;
+      if (!results) return false;
+      for (const result of results) {
+        if (result.sourceShape.name === 'at-least-one-article-container') {
+          missingArticleBlock = true;
+        }
       }
-    );
+      return missingArticleBlock;
+    } else {
+      return false;
+    }
   }
 }