+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- validate motivering and article section
+=======
+
 ## [7.0.0] - 2023-05-17
 ### Added
 - add docker build for easy demo environments
@@ -14,9 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BREAKING: Endpoint config for `CitationPlugin`
 ### Fixed
 - Remove structure doesn't always work - disable the "Remove <structure>" button correctly
+- Fix initialization of default date properties
 
 ### Fixed
-- Fix initialization of default date properties
+
 ## [6.1.0] - 2023-05-11
 
 ### Added